Move struct/classes into seperate files
