+++ Sources/Request Strategies/Conversation/ConversationRequestStrategy.swift
@@ -17,96 +17,6 @@
 
 import Foundation
 
-protocol Paginatable: Decodable {
-    var hasMore: Bool { get }
-    var nextStartReference: String? { get }
-}
-
-protocol PaginatableTranscoder: class {
-
-    associatedtype Result: Paginatable
-
-    func didReceiveResults(results: Result)
-    func didPermanent(results: Result)
-
-}
-
-class PaginatedSync<Payload: Paginatable>: NSObject, ZMRequestGenerator {
-
-    typealias CompletionHandler = (Swift.Result<Payload, PaginatedSyncError>) -> Void
-
-    enum Status: Equatable {
-        case fetching(_ state: String)
-        case done
-    }
-
-    enum PaginatedSyncError: Error {
-        case permanentError
-    }
-
-    let context: NSManagedObjectContext
-    let basePath: String
-    let pageSize: Int
-    var status: Status = .done
-    var request: ZMTransportRequest? = nil
-    var completionHandler: CompletionHandler?
-
-    init(basePath: String, pageSize: Int, context: NSManagedObjectContext) {
-        self.basePath = basePath
-        self.pageSize = pageSize
-        self.context = context
-    }
-
-    func fetch(_ completionHandler: @escaping CompletionHandler) {
-        self.completionHandler = completionHandler
-        status = .fetching("")
-    }
-
-    func nextRequest() -> ZMTransportRequest? {
-        guard request == nil, case .fetching(let start) = status else {
-            return nil
-        }
-
-        var queryItems = [URLQueryItem(name: "size", value: String(pageSize))]
-
-        if !start.isEmpty {
-            queryItems.append(URLQueryItem(name: "start", value: start))
-        }
-
-        var urlComponents = URLComponents(string: basePath)
-        urlComponents?.queryItems = queryItems
-
-        guard let path = urlComponents?.string else {
-            return nil
-        }
-
-        self.request = ZMTransportRequest(getFromPath: path)
-
-        request?.add(ZMCompletionHandler(on: context, block: { (response) in
-            self.request = nil
-
-            guard let result = Payload(response, decoder: .defaultDecoder) else {
-                if response.result == .permanentError {
-                    self.completionHandler?(.failure(.permanentError))
-                    self.status = .done
-                }
-                return
-            }
-
-            self.completionHandler?(.success(result))
-
-            if result.hasMore, let nextStartReference = result.nextStartReference {
-                self.status = .fetching(nextStartReference)
-            } else {
-                self.status = .done
-            }
-        }))
-
-        return request
-    }
-
-}
-
 public class ConversationRequestStrategy: AbstractRequestStrategy, ZMRequestGeneratorSource, ZMContextChangeTrackerSource {
 
     let syncProgress: SyncProgress
