+++ Sources/Request Syncs/PaginatedSync.swift
@@ -1,0 +1,100 @@
+//
+// Wire
+// Copyright (C) 2021 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import Foundation
+
+protocol Paginatable: Decodable {
+    var hasMore: Bool { get }
+    var nextStartReference: String? { get }
+}
+
+class PaginatedSync<Payload: Paginatable>: NSObject, ZMRequestGenerator {
+
+    typealias CompletionHandler = (Swift.Result<Payload, PaginatedSyncError>) -> Void
+
+    enum Status: Equatable {
+        case fetching(_ state: String)
+        case done
+    }
+
+    enum PaginatedSyncError: Error {
+        case permanentError
+    }
+
+    let context: NSManagedObjectContext
+    let basePath: String
+    let pageSize: Int
+    var status: Status = .done
+    var request: ZMTransportRequest? = nil
+    var completionHandler: CompletionHandler?
+
+    init(basePath: String, pageSize: Int, context: NSManagedObjectContext) {
+        self.basePath = basePath
+        self.pageSize = pageSize
+        self.context = context
+    }
+
+    func fetch(_ completionHandler: @escaping CompletionHandler) {
+        self.completionHandler = completionHandler
+        status = .fetching("")
+    }
+
+    func nextRequest() -> ZMTransportRequest? {
+        guard request == nil, case .fetching(let start) = status else {
+            return nil
+        }
+
+        var queryItems = [URLQueryItem(name: "size", value: String(pageSize))]
+
+        if !start.isEmpty {
+            queryItems.append(URLQueryItem(name: "start", value: start))
+        }
+
+        var urlComponents = URLComponents(string: basePath)
+        urlComponents?.queryItems = queryItems
+
+        guard let path = urlComponents?.string else {
+            return nil
+        }
+
+        self.request = ZMTransportRequest(getFromPath: path)
+
+        request?.add(ZMCompletionHandler(on: context, block: { (response) in
+            self.request = nil
+
+            guard let result = Payload(response, decoder: .defaultDecoder) else {
+                if response.result == .permanentError {
+                    self.completionHandler?(.failure(.permanentError))
+                    self.status = .done
+                }
+                return
+            }
+
+            self.completionHandler?(.success(result))
+
+            if result.hasMore, let nextStartReference = result.nextStartReference {
+                self.status = .fetching(nextStartReference)
+            } else {
+                self.status = .done
+            }
+        }))
+
+        return request
+    }
+
+}
